feat: enhance projections management with improved delete handling and UI updates

+++ lib/vantage_web/live/projection_live/index.ex
@@ -1,4 +1,5 @@
 defmodule VantageWeb.ProjectionLive.Index do
+  require Logger
   use VantageWeb, :live_view
 
   alias Vantage.Projections
@@ -41,16 +42,29 @@ defmodule VantageWeb.ProjectionLive.Index do
   end
 
   @impl true
-  def handle_event("delete", %{"id" => id, "file" => file}, socket) do
-    file_path = Path.join([:code.priv_dir(:vantage), "static", "uploads", Path.basename(file)])
+  def handle_event(
+        "delete",
+        %{"id" => id, "file" => file, "investigation_id" => investigation_id},
+        socket
+      ) do
+    if file do
+      file_path = Path.join([:code.priv_dir(:vantage), "static", "uploads", Path.basename(file)])
 
-    if File.exists?(file_path) do
-      File.rm!(file_path)
+      if File.exists?(file_path) do
+        File.rm!(file_path)
+      end
     end
 
     projection = Projections.get_projection!(id)
     {:ok, _} = Projections.delete_projection(projection)
 
-    {:noreply, stream_delete(socket, :projections, projection)}
+    # Logger.warning(~p"/investigations/#{investigation_id}")
+
+    {
+      :noreply,
+      socket
+      |> stream_delete(:projections, projection)
+      #  |> push_patch(to: ~p"/investigations/#{investigation_id}/projections", replace: true)
+    }
   end
 end
